Remove CJS from shared.client
