+++ .changeset/happy-cycles-glow.md
@@ -1,0 +1,5 @@
+---
+"@osdk/shared.client": patch
+---
+
+Removed CJS support
