+++ packages/e2e.sandbox.oauth.public.react-router/vite.config.ts
@@ -7,7 +7,4 @@ export default defineConfig({
   server: {
     port: 8080,
   },
-  optimizeDeps: {
-    include: ["@osdk/shared.client"],
-  },
 });
