+++ packages/shared.client/package.json
@@ -8,12 +8,11 @@
   },
   "exports": {
     ".": {
-      "require": "./index.cjs",
       "import": "./index.mjs"
     }
   },
   "scripts": {
-    "check-attw": "monorepo.tool.attw both",
+    "check-attw": "monorepo.tool.attw esm",
     "clean": "rm -rf lib dist types build tsconfig.tsbuildinfo",
     "fix-lint": "eslint . --fix && dprint fmt --config $(find-up dprint.json)",
     "lint": "eslint . && dprint check  --config $(find-up dprint.json)"
@@ -35,7 +34,6 @@
     "index.cjs",
     "index.mjs"
   ],
-  "main": "index.cjs",
   "module": "index.mjs",
   "types": "index.d.mts",
   "type": "module"
